backend/requirements.txt
django==4.0.2
Pillow==9.0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
django-cors-headers